# api/openapi.yaml
openapi: 3.1.0
info:
  title: Touch API
  version: 1.0.0
  description: Standalone phone and watch touch delivery API with optional durable voice attachments.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/installations/{installationId}:
    put:
      operationId: registerInstallation
      parameters:
        - $ref: '#/components/parameters/InstallationId'
        - $ref: '#/components/parameters/InstallationHeader'
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/InstallationRequest' } }
      responses:
        '200': { description: Installation registered, content: { application/json: { schema: { $ref: '#/components/schemas/Installation' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
    get:
      operationId: getInstallationStatus
      summary: Verify the authenticated installation, username, and server-side FCM token fingerprint.
      parameters:
        - $ref: '#/components/parameters/InstallationId'
        - $ref: '#/components/parameters/InstallationHeader'
      responses:
        '200': { description: Installation status, content: { application/json: { schema: { $ref: '#/components/schemas/InstallationStatus' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/installations/{installationId}/push-tests:
    post:
      operationId: startPushTest
      parameters: [ { $ref: '#/components/parameters/InstallationId' }, { $ref: '#/components/parameters/InstallationHeader' } ]
      responses:
        '202': { description: Push challenge created, content: { application/json: { schema: { $ref: '#/components/schemas/PushTest' } } } }
        '401': { $ref: '#/components/responses/Error' }
  /v1/push-tests/{testId}:
    get:
      operationId: getPushTest
      parameters: [ { name: testId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }, { $ref: '#/components/parameters/InstallationHeader' } ]
      responses:
        '200': { description: Current challenge result, content: { application/json: { schema: { $ref: '#/components/schemas/PushTest' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /v1/push-tests/{testId}/ack:
    post:
      operationId: acknowledgePushTest
      parameters: [ { name: testId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }, { $ref: '#/components/parameters/InstallationHeader' } ]
      responses:
        '200': { description: Challenge delivered, content: { application/json: { schema: { $ref: '#/components/schemas/PushTest' } } } }
  /v1/contacts:
    get:
      operationId: listContacts
      parameters: [ { $ref: '#/components/parameters/InstallationHeader' } ]
      responses:
        '200': { description: Saved and recent contacts, content: { application/json: { schema: { type: object, required: [contacts], properties: { contacts: { type: array, items: { $ref: '#/components/schemas/Contact' } } } } } } }
    post:
      operationId: addContact
      parameters: [ { $ref: '#/components/parameters/InstallationHeader' } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [username], properties: { username: { $ref: '#/components/schemas/Username' } } } } } }
      responses:
        '201': { description: Contact saved, content: { application/json: { schema: { $ref: '#/components/schemas/Contact' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /v1/contacts/{username}:
    delete:
      operationId: removeContact
      parameters: [ { name: username, in: path, required: true, schema: { $ref: '#/components/schemas/Username' } }, { $ref: '#/components/parameters/InstallationHeader' } ]
      responses: { '204': { description: Contact removed from saved contacts. } }
  /v1/diagnostics/events:
    post:
      operationId: uploadDiagnostics
      parameters: [ { $ref: '#/components/parameters/InstallationHeader' } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [events], properties: { events: { type: array, minItems: 1, maxItems: 100, items: { $ref: '#/components/schemas/DiagnosticEvent' } } } } } } }
      responses: { '204': { description: Events accepted. }, '400': { $ref: '#/components/responses/Error' } }
...
  /v1/touches:
    post:
      operationId: sendTouch
      parameters: [ { $ref: '#/components/parameters/InstallationHeader' } ]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/TouchSendRequest' } }
      responses:
        '202': { description: Touch durably accepted, content: { application/json: { schema: { $ref: '#/components/schemas/TouchAccepted' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
    get:
      operationId: listPendingTouches
      parameters:
        - $ref: '#/components/parameters/InstallationHeader'
        - name: state
          in: query
          required: true
          schema: { type: string, const: pending }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        '200': { description: FIFO pending inbox, content: { application/json: { schema: { $ref: '#/components/schemas/PendingTouches' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/touches/{touchId}/ack:
    post:
      operationId: acknowledgeTouch
      parameters:
        - $ref: '#/components/parameters/TouchId'
        - $ref: '#/components/parameters/InstallationHeader'
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/AckRequest' } }
      responses:
        '204': { description: Acknowledgement recorded. }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/touches/{touchId}/status:
    get:
      operationId: getTouchDeliveryStatus
      parameters: [ { $ref: '#/components/parameters/TouchId' }, { $ref: '#/components/parameters/InstallationHeader' } ]
      responses:
        '200': { description: Sender-visible delivery status, content: { application/json: { schema: { $ref: '#/components/schemas/TouchDeliveryStatus' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /v1/admin/diagnostics:
    get:
      operationId: adminDiagnostics
      security: [ { adminBearer: [] } ]
      parameters:
        - { name: callId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: installationId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: messageId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: severity, in: query, schema: { type: string } }
        - { name: sinceMs, in: query, schema: { type: integer, format: int64 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses: { '200': { description: Recent structured client diagnostics. }, '401': { $ref: '#/components/responses/Error' } }
...
  /v1/admin/calls:
    get:
      operationId: adminCalls
      security: [ { adminBearer: [] } ]
      responses: { '200': { description: Recent call sessions and transport counters. }, '401': { $ref: '#/components/responses/Error' } }
  /v1/admin/clients:
    get:
      operationId: adminClients
      security: [ { adminBearer: [] } ]
      responses: { '200': { description: Registered client summaries plus current ephemeral Live socket and stream state; no FCM tokens or media are returned. }, '401': { $ref: '#/components/responses/Error' } }
  /v1/admin/calls/{callId}:
    get:
      operationId: adminCallTimeline
      security: [ { adminBearer: [] } ]
      parameters: [ { name: callId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } } ]
      responses: { '200': { description: Ordered server and client call-state timeline. }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' } }
  /v1/admin/notifications:
    get:
      operationId: adminNotifications
      security: [ { adminBearer: [] } ]
      responses: { '200': { description: Recent FCM attempts and client delivery events. }, '401': { $ref: '#/components/responses/Error' } }
...
components:
  securitySchemes:
    adminBearer: { type: http, scheme: bearer }
  parameters:
    InstallationId:
      name: installationId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    TouchId:
      name: touchId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    InstallationHeader:
      name: X-Installation-ID
      in: header
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
  responses:
    Error:
      description: Stable API error.
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }
  schemas:
    Uuid: { type: string, format: uuid }
    Username: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_.-]{2,31}$' }
...
    InstallationRequest:
      type: object
      additionalProperties: false
      required: [username, platform]
      properties:
        username: { $ref: '#/components/schemas/Username' }
        platform: { type: string, enum: [phone, watch] }
        fcmToken: { type: string }
    Installation:
      type: object
      required: [installationId, username, platform, updatedAtMs]
      properties:
        installationId: { $ref: '#/components/schemas/Uuid' }
        username: { $ref: '#/components/schemas/Username' }
        platform: { type: string, enum: [phone, watch] }
        updatedAtMs: { type: integer, format: int64 }
    InstallationStatus:
      allOf:
        - { $ref: '#/components/schemas/Installation' }
        - type: object
          required: [fcmTokenPresent]
          properties:
            fcmTokenPresent: { type: boolean }
            fcmTokenFingerprint: { type: string, description: First 12 hexadecimal characters of SHA-256; never the raw token. }
            lastPushTestStatus: { type: string }
    Contact:
      type: object
      required: [username, saved]
      properties:
        username: { $ref: '#/components/schemas/Username' }
        saved: { type: boolean }
        lastUsedAtMs: { type: integer, format: int64 }
    PushTest:
      type: object
      required: [testId, status, createdAtMs]
      properties:
        testId: { $ref: '#/components/schemas/Uuid' }
        status: { type: string, enum: [pending, delivered, timed_out, invalid_token, service_error] }
        createdAtMs: { type: integer, format: int64 }
        completedAtMs: { type: integer, format: int64 }
    DiagnosticEvent:
      type: object
      additionalProperties: false
      required: [eventId, occurredAtMs, severity, category, name]
      properties:
        eventId: { $ref: '#/components/schemas/Uuid' }
        occurredAtMs: { type: integer, format: int64 }
        severity: { type: string, enum: [debug, info, warn, error] }
        category: { type: string, maxLength: 32 }
        name: { type: string, maxLength: 64 }
        callId: { $ref: '#/components/schemas/Uuid' }
        touchId: { $ref: '#/components/schemas/Uuid' }
        clientMessageId: { $ref: '#/components/schemas/Uuid' }
        message: { type: string, maxLength: 512 }
        attributes: { type: object, maxProperties: 32 }
...
    TouchDeliveryStatus:
      type: object
      required: [touchId, acceptedAtMs, notificationStatus]
      properties:
        touchId: { $ref: '#/components/schemas/Uuid' }
        acceptedAtMs: { type: integer, format: int64 }
        notificationStatus: { type: string }
        notificationReceivedAtMs: { type: integer, format: int64 }
        persistedAtMs: { type: integer, format: int64 }
        playedAtMs: { type: integer, format: int64 }
    TouchSendRequest:
      type: object
      additionalProperties: false
      required: [clientMessageId, recipientUsername, samplePeriodMs, amplitudes]
      properties:
        clientMessageId: { $ref: '#/components/schemas/Uuid' }
        recipientUsername: { $ref: '#/components/schemas/Username' }
        samplePeriodMs: { type: integer, minimum: 1, maximum: 100 }
        amplitudes: { type: array, minItems: 1, maxItems: 3000, items: { type: integer, minimum: 0, maximum: 255 } }
        audio: { $ref: '#/components/schemas/AudioAttachment' }
    TouchAccepted:
      type: object
      required: [touchId, clientMessageId, acceptedAtMs, duplicate]
      properties:
        touchId: { $ref: '#/components/schemas/Uuid' }
        clientMessageId: { $ref: '#/components/schemas/Uuid' }
        acceptedAtMs: { type: integer, format: int64 }
        duplicate: { type: boolean }
    PendingTouch:
      type: object
      required: [touchId, clientMessageId, senderUsername, samplePeriodMs, amplitudes, createdAtMs, expiresAtMs]
      properties:
        touchId: { $ref: '#/components/schemas/Uuid' }
        clientMessageId: { $ref: '#/components/schemas/Uuid' }
        senderUsername: { $ref: '#/components/schemas/Username' }
        samplePeriodMs: { type: integer }
        amplitudes: { type: array, items: { type: integer } }
        createdAtMs: { type: integer, format: int64 }
        expiresAtMs: { type: integer, format: int64 }
        audio: { $ref: '#/components/schemas/AudioAttachment' }
    PendingTouches:
      type: object
      required: [touches]
      properties: { touches: { type: array, items: { $ref: '#/components/schemas/PendingTouch' } } }
    AckRequest:
      type: object
      additionalProperties: false
      required: [state]
      properties: { state: { type: string, enum: [persisted, played] } }
...
    AudioAttachment:
      type: object
      additionalProperties: false
      required: [codec, sampleRateHz, channelCount, durationMs, data]
      properties:
        codec: { type: string, const: aac-lc }
        sampleRateHz: { type: integer, const: 16000 }
        channelCount: { type: integer, const: 1 }
        durationMs: { type: integer, minimum: 1, maximum: 30000 }
        data: { type: string, format: byte, description: Base64 MP4/AAC-LC data, maximum decoded size 256 KiB. }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }